+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Benchmarks/ItemStreamBenchmark.cs
@@ -1,0 +1,172 @@
+﻿// ----------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// ----------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos.Performance.Tests.Benchmarks
+{
+    using System;
+    using System.IO;
+    using System.Linq;
+    using System.Net;
+    using System.Threading.Tasks;
+    using BenchmarkDotNet.Attributes;
+    using Microsoft.Azure.Cosmos;
+
+    public class ItemStreamBenchmark : IItemBenchmark
+    {
+        private ItemBenchmarkHelper benchmarkHelper;
+
+        public ItemStreamBenchmark()
+        {
+            this.benchmarkHelper = new ItemBenchmarkHelper();
+        }
+
+        /// <summary>
+        /// Benchmark for CreateItemAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task CreateItem()
+        {
+            using (MemoryStream ms = new MemoryStream(this.benchmarkHelper.TestItemBytes))
+            {
+                using (ResponseMessage response = await this.benchmarkHelper.TestContainer.CreateItemStreamAsync(
+                    ms,
+                    new Cosmos.PartitionKey(Constants.ValidOperationId)))
+                {
+                    if ((int)response.StatusCode > 300 || response.Content == null)
+                    {
+                        throw new Exception();
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Benchmark for UpsertItemAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task UpsertItem()
+        {
+            using (ResponseMessage response = await this.benchmarkHelper.TestContainer.UpsertItemStreamAsync(
+                new MemoryStream(this.benchmarkHelper.TestItemBytes),
+                new Cosmos.PartitionKey(Constants.ValidOperationId)))
+            {
+                if ((int)response.StatusCode > 300 || response.Content == null)
+                {
+                    throw new Exception();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Benchmark for ReadItemAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task ReadItemNotExists()
+        {
+            using (ResponseMessage response = await this.benchmarkHelper.TestContainer.ReadItemStreamAsync(
+                Constants.NotFoundOperationId,
+                new Cosmos.PartitionKey(Constants.ValidOperationId)))
+            {
+                if (response.StatusCode != System.Net.HttpStatusCode.NotFound)
+                {
+                    throw new Exception();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Benchmark for ReadItemStreamAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task ReadItemExists()
+        {
+            using (ResponseMessage response = await this.benchmarkHelper.TestContainer.ReadItemStreamAsync(
+                Constants.ValidOperationId,
+                new Cosmos.PartitionKey(Constants.ValidOperationId)))
+            {
+                if (response.StatusCode == System.Net.HttpStatusCode.NotFound || response.Content == null)
+                {
+                    throw new Exception();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Benchmark for ReplaceItemAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task UpdateItem()
+        {
+            using (ResponseMessage response = await this.benchmarkHelper.TestContainer.ReplaceItemStreamAsync(
+                new MemoryStream(this.benchmarkHelper.TestItemBytes),
+                Constants.ValidOperationId,
+                new Cosmos.PartitionKey(Constants.ValidOperationId)))
+            {
+                if (response.StatusCode == System.Net.HttpStatusCode.NotFound || response.Content == null)
+                {
+                    throw new Exception();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Benchmark for DeleteItemAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task DeleteItemExists()
+        {
+            using (ResponseMessage response = await this.benchmarkHelper.TestContainer.DeleteItemStreamAsync(
+                Constants.ValidOperationId,
+                new Cosmos.PartitionKey(Constants.ValidOperationId)))
+            {
+                if (response.StatusCode == System.Net.HttpStatusCode.NotFound)
+                {
+                    throw new Exception();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Benchmark for DeleteItemAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task DeleteItemNotExists()
+        {
+            using (ResponseMessage response = await this.benchmarkHelper.TestContainer.DeleteItemStreamAsync(
+                Constants.NotFoundOperationId,
+                new Cosmos.PartitionKey(Constants.ValidOperationId)))
+            {
+                if (response.StatusCode != System.Net.HttpStatusCode.NotFound)
+                {
+                    throw new Exception();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Benchmark for DeleteItemAsync.
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        [Benchmark]
+        public async Task ReadFeed()
+        {
+            FeedIterator streamIterator = this.benchmarkHelper.TestContainer.GetItemQueryStreamIterator();
+            while (streamIterator.HasMoreResults)
+            {
+                ResponseMessage response = await streamIterator.ReadNextAsync();
+                if (response.StatusCode != HttpStatusCode.OK)
+                {
+                    throw new Exception();
+                }
+            }
+        }
+    }
+}
